feat: move keys suites
